account for split/subcomm not being moved. forgot one.
