+++ libensemble/tests/regression_tests/test_chwirut_pounders_subcomm.py
@@ -20,7 +20,8 @@ from libensemble.libE import libE
 from libensemble.sim_funcs.chwirut1 import chwirut_eval as sim_f
 from libensemble.gen_funcs.aposmm import aposmm_logic as gen_f
 from libensemble.tests.regression_tests.support import persis_info_2 as persis_info, aposmm_gen_out as gen_out
-from libensemble.utils import parse_args, save_libE_output, per_worker_stream, mpi_comm_excl
+from libEnsemble.tests.regression_tests.common import mpi_comm_excl
+from libensemble.utils import parse_args, save_libE_output, per_worker_stream
 
 nworkers, is_master, libE_specs, _ = parse_args()
 libE_specs['comm'], mpi_comm_null = mpi_comm_excl()
